requirements.txt
pandas==2.1.0
numpy==1.24.3
pytest==7.4.0
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pyyaml==6.0.1
black==23.7.0
mypy==1.5.1
pylint==2.17.5
aiohttp==3.8.5
pyarrow==13.0.0 